+++ app/ante/ante.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"runtime/debug"
 
-	"github.com/palantir/stacktrace"
-
 	tmlog "github.com/tendermint/tendermint/libs/log"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -68,9 +66,10 @@ func NewAnteHandler(
 					)
 
 				default:
-					return ctx, stacktrace.Propagate(
-						sdkerrors.Wrap(sdkerrors.ErrUnknownExtensionOptions, typeURL),
-						"rejecting tx with unsupported extension option",
+					return ctx, sdkerrors.Wrapf(
+						sdkerrors.ErrUnknownExtensionOptions,
+						"rejecting tx with unsupported extension option: %s",
+						typeURL,
 					)
 				}
 
@@ -99,10 +98,7 @@ func NewAnteHandler(
 				authante.NewIncrementSequenceDecorator(ak), // innermost AnteDecorator
 			)
 		default:
-			return ctx, stacktrace.Propagate(
-				sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "invalid transaction type: %T", tx),
-				"transaction is not an SDK tx",
-			)
+			return ctx, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "invalid transaction type: %T", tx)
 		}
 
 		return anteHandler(ctx, tx, sim)